feat(multiple): integrated the refactor/rewrite that ends up cleaning up a couple of issues
this change integrates an extraction of several parts of the original plugin to separate modules to

improve the overall design and facilitate better testability. as part of the integrated changes, we

now check the rolled-up commit statuses rather than looping over the individual statuses. config is

now accepted as plugin options rather than environment variables, which is also slightly simplified

because the github user is no longer needed as a result of switching to token auth.

BREAKING CHANGES: config is now expected as options to the hapi plugin rather than environment

variables. it is up to the consumer how they want to gather the config, but they must now be passed

to the plugin as options. closes #3. closes #4. closes #5. closes #17. closes #20