feat(status-event): enabled processing to be triggered by a status-event
hoping that use of the status-event can replace the use of the pull_request-event, but this enables
the choice for now. the intent is that one or the other should be used, not both. this is to prevent
collisions between polling from the pull_request-event and the more intentional status-event
processing. switching to the status-event eliminates polling, which allows acceptance to be
triggered after what would have been beyond the polling timeout. hopefully this will also improve
the behavior related to #77 because a successful status would need to be reported after the PR is
opened. ideally this gives enough time for all checks to be registered

for #21